+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added `nn.Path.Paths()` method
 - added `nn.VarStore.Summary()` method
 - fixed incorrect tensor method `ts.Meshgrid` -> `Meshgrid`
+- added new API `ConstantPadNdWithVal` `ato_constant_pad_nd` with padding value.
 
 ## [Nofix]
 - ctype `long` caused compiling error in MacOS as noted on [#44]. Not working on linux box.